clean, break out into separate components
