+++ apps/web/src/modules/dao/components/AuctionChart/index.ts
@@ -1,0 +1,1 @@
+export * from './AuctionChart'
